Module Dirsp_exchange

For securing a conversation between two or more parties.

Dirsp is a short form for Diskuv Implementations of Research Security Protocols. The dirsp-exchange library has implementations of protocols published in the research field of the authenticated exchange of messages. It uses a dirsp-proscript library like dirsp-proscript-mirage to provide a runtime library of cryptographic primitives.

At the moment this library is a no-op. The dirsp-exchange-kbb2017 library is the first implementation of a secure two party protocol.